# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateNodeRequest": {
        "description": "节点注册请求：兼容两种格式\n- 面板/测试：`{\"node\": {name, hostname, ip_address, status, ...}}`\n- Agent 平铺：`{name, host, agent_port, auth_token}`",
        "properties": {
          "agent_port": {
            "description": "Agent HTTP 服务端口（默认 9527）",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "auth_token": {
            "type": [
              "string",
              "null"
            ]
          },
          "hostname": {
            "type": "string"
          },
          "ip_address": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "node": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ServerNode"
              }
            ]
          },
          "status": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
      "HeartbeatRequest": {
        "properties": {
          "cpu_usage": {
            "format": "float",
            "type": "number"
          },
          "disk_usage_percent": {
            "format": "float",
            "type": "number"
          },
          "load_one": {
            "format": "float",
            "type": "number"
          },
          "memory_usage_percent": {
            "format": "float",
            "type": "number"
          }
        },
        "required": [
          "cpu_usage",
          "memory_usage_percent",
          "disk_usage_percent",
          "load_one"
        ],
        "type": "object"
      },
...
      "PaginatedResponse<ServerNode>_ServerNode": {
        "properties": {
          "data": {
            "items": {
              "properties": {
                "agent_port": {
                  "description": "Agent HTTP 服务端口（面板据此构造 `http://<ip>:<agent_port>` 发起远程调用）",
                  "format": "int32",
                  "minimum": 0,
                  "type": "integer"
                },
                "auth_token": {
                  "description": "Agent 注册时携带的认证令牌",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "created_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "hostname": {
                  "type": "string"
                },
                "id": {
                  "format": "int64",
                  "type": "integer"
                },
                "ip_address": {
                  "type": "string"
                },
                "last_heartbeat_at": {
                  "description": "最近一次心跳时间（Agent 上报）",
                  "format": "date-time",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "metrics_json": {
                  "description": "最近指标快照 JSON：{cpu_usage, memory_usage_percent, disk_usage_percent, load_one}",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "name": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "hostname",
                "ip_address",
                "status",
                "created_at"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          }
        },
        "required": [
          "data",
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "type": "object"
      },
...
      "RemoteActionRequest": {
        "description": "Phase A1：Agent 动作枚举请求（动作 + 参数）",
        "properties": {
          "action": {
            "type": "string"
          },
          "params": {}
        },
        "required": [
          "action"
        ],
        "type": "object"
      },
      "RemoteBatchExecRequest": {
        "properties": {
          "command": {
            "type": "string"
          },
          "node_ids": {
            "items": {
              "format": "int64",
              "type": "integer"
            },
            "type": "array"
          },
          "timeout_secs": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "node_ids",
          "command"
        ],
        "type": "object"
      },
      "RemoteExecRequest": {
        "properties": {
          "command": {
            "type": "string"
          },
          "timeout_secs": {
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          }
        },
        "required": [
          "command"
        ],
        "type": "object"
      },
...
      "RemoteUploadRequest": {
        "properties": {
          "content_base64": {
            "description": "base64 编码的文件内容",
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        },
        "required": [
          "path",
          "content_base64"
        ],
        "type": "object"
      },
...
      "ServerNode": {
        "properties": {
          "agent_port": {
            "description": "Agent HTTP 服务端口（面板据此构造 `http://<ip>:<agent_port>` 发起远程调用）",
            "format": "int32",
            "minimum": 0,
            "type": "integer"
          },
          "auth_token": {
            "description": "Agent 注册时携带的认证令牌",
            "type": [
              "string",
              "null"
            ]
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "hostname": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "ip_address": {
            "type": "string"
          },
          "last_heartbeat_at": {
            "description": "最近一次心跳时间（Agent 上报）",
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "metrics_json": {
            "description": "最近指标快照 JSON：{cpu_usage, memory_usage_percent, disk_usage_percent, load_one}",
            "type": [
              "string",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "hostname",
          "ip_address",
          "status",
          "created_at"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "FlamePanel 面板后端 REST API（OpenAPI 3.x，由 utoipa 编译期生成）",
    "license": {
      "name": ""
    },
    "title": "FlamePanel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/nodes": {
      "get": {
        "operationId": "list_nodes",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int64",
              "minimum": 1,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "每页条数（1~200）",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int64",
              "maximum": 200,
              "minimum": 1,
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse<ServerNode>_ServerNode"
                }
              }
            },
            "description": "节点列表"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "节点列表（分页）",
        "tags": [
          "nodes"
        ]
      },
      "post": {
        "operationId": "create_node",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "format": "int64",
                  "type": "integer"
                }
              }
            },
            "description": "注册成功，返回节点 ID"
          },
          "400": {
            "description": "参数错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "注册节点",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/batch-execute": {
      "post": {
        "operationId": "batch_execute",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoteBatchExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "各节点执行结果聚合"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "批量在多个节点并行执行命令",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/heartbeat/{id}": {
      "post": {
        "operationId": "heartbeat",
        "parameters": [
          {
            "description": "节点 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HeartbeatRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "心跳记录成功"
          },
          "401": {
            "description": "Agent token 无效"
          }
        },
        "summary": "Agent 心跳上报（白名单免 JWT，校验 Agent token）",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/register": {
      "post": {
        "operationId": "register_agent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNodeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "注册成功，返回节点 ID"
          },
          "400": {
            "description": "参数错误"
          }
        },
        "summary": "Agent 注册（白名单免 JWT，供 Agent 启动时自动注册）\n请求体：`{name, host, agent_port, auth_token}`，返回 `{\"id\": <node_id>}`",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/{id}/action": {
      "post": {
        "operationId": "remote_action",
        "parameters": [
          {
            "description": "节点 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoteActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "动作结果"
          },
          "400": {
            "description": "非法动作或参数"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "调用 Agent 动作枚举（Phase A1：安全动作白名单）",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/{id}/execute": {
      "post": {
        "operationId": "remote_execute",
        "parameters": [
          {
            "description": "节点 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoteExecRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "执行结果"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "在远程节点执行命令",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/{id}/files": {
      "get": {
        "operationId": "remote_list_files",
        "parameters": [
          {
            "description": "节点 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "远程目录列表"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "列出远程节点目录",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/{id}/files/download": {
      "get": {
        "operationId": "remote_download_file",
        "parameters": [
          {
            "description": "节点 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "文件内容（base64）"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "下载远程节点文件",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/{id}/files/upload": {
      "post": {
        "operationId": "remote_upload_file",
        "parameters": [
          {
            "description": "节点 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoteUploadRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "上传结果"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "上传文件到远程节点",
        "tags": [
          "nodes"
        ]
      }
    },
    "/api/nodes/{id}/status": {
      "get": {
        "operationId": "status",
        "parameters": [
          {
            "description": "节点 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "在线状态"
          },
          "404": {
            "description": "节点不存在"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "节点在线状态（惰性判定）",
        "tags": [
          "nodes"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "默认",
      "url": "/"
    }
  ]
}